# docs/swagger.yaml
basePath: /
definitions:
...
  common.jsonResult:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact:
    name: OpenList Team
    url: https://github.com/OpenListTeam/
  description: A file list program that supports multiple storage.
  license:
    name: GNU Affero General Public License v3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  termsOfService: https://doc.oplist.org/terms
  title: OpenList
  version: latest
paths:
...
  /api/public/archive_extensions:
    get:
      consumes:
      - application/json
      description: Get a list of supported archive extensions
      produces:
      - application/json
      responses:
        "200":
          description: List of archive extensions
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/common.jsonResult'
            - properties:
                data:
                  type: string
              type: object
      summary: Get supported archive extensions
      tags:
      - Public
...
securityDefinitions:
  Authorization:
    description: Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"